NeighborlyAPI/requirements.txt
azure-functions==1.2.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==2.2.1
dominate==2.5.1
feedgen==0.9.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.9.1
MarkupSafe==1.1.1
pymongo==4.3.2
python-dateutil==2.8.2
requests==2.23.0
six==1.16.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==0.16.1
